updating productDao so when user requests product it gets a double for price instead of int and updating database table to use decimal column instead of number

+++ src/main/java/Shop/product/ProductDao.java
@@ -32,7 +32,7 @@ public class ProductDao {
             preparedStatement.setString(1, "%" + productId + "%");
             ResultSet result = preparedStatement.executeQuery();
             while (result.next()) {
-                Product product = new Product(result.getString("product_name"), result.getInt("stock"), result.getInt("price"));
+                Product product = new Product(result.getString("product_name"), result.getInt("stock"), result.getDouble("price"));
                 list.add(product);
             }
         } catch (SQLException e) {
